Added gdbserver support
To debug the simulator uncomment the following line in
simulation_control/run.py:
GDB_OPTIONS = ' --gdb-server'

Then start the simulation as usual. It should stop and wait for the
debugger to attach.
If you get the following error:

Error: Unable to enable debugging via gdb when the yama ptrace_scope protection is enabled.
 Disable ptrace_scope by setting /proc/sys/kernel/yama/ptrace_scope to 0
 or by updating the value in /etc/sysctl.d/10-ptrace.conf.
 === GDB permission error ===

set the ptrace_scope to 0 with the command on your host:

echo 0 | sudo tee /proc/sys/kernel/yama/ptrace_scope

Now you can start a gdb session outside the docker with:
$ gdb
or
$ gdb <path to HotSniper/lib/sniper>

Then attach to the gdbserver with the gdb command:
(gdb) target remote localhost:7000

You can now set a breakpoint (eg. in 'main' or 'SchedulerOpen::SchedulerOpen').
The 'continue' command will run the simulator until the breakpoint is
hit.